+++ packages/date-publication/.eslintrc.json
@@ -1,3 +1,6 @@
 {
-  "extends": ["@times-components/thetimes"]
+  "extends": ["@times-components/thetimes"],
+  "env": {
+    "jest": true
+  }
 }
