+++ packages/date-publication/__mocks__/react-native-device-info.js
