+++ packages/date-publication/__tests__/ios/date-publication.ios.test.js
@@ -1,6 +1,4 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
 import rndiMock from "react-native-device-info";
-
 import shared from "../shared";
 
 describe("Date Publication test on iOS when the user has the same time zone as London time zone", () => {
